actualizacion de vida desde ui unity

+++ Assets/Scripts/architecture_business/use_cases/player_spacecraft/life/SpacecraftPlayerQuitLifeUseCase.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 public interface SpacecraftPlayerQuitLifeUseCase 
 {
-    public void invoke(float life);
+    public void invoke(int life);
 }
 
 public class SpacecraftPlayerQuitLifeUseCaseImpl : SpacecraftPlayerQuitLifeUseCase
 {
     private SpacecraftPlayerLifeRepository repo = new SpacecraftPlayerLifeRepositoryImpl();
-    public void invoke(float life) => repo.quitLife(life);
+    public void invoke(int life) => repo.quitLife(life);
 }
